Phase 3.1aa fix round 1: review feedback (P1, P2)
P1 (freshness):
- Repository upsert now uses two-gate change detection. fetched_at
  advances on successful live re-fetch even when content is byte-equal,
  preventing refresh_all_stale from re-selecting the same row forever.
  True no-op only when content + fetched_at are both identical.
- New repo test: identical content + advanced fetched_at returns
  changed=True and updates the row.

P2 (cache-key integrity):
- New CubeMetadataProductMismatchError raised by get_or_fetch and
  refresh when cached row's product_id differs from the caller's
  argument. No auto-heal — exposes upstream cube_id↔product_id
  mapping bugs instead of silently masking them. StatCan client is
  not called when the mismatch is detected.
- Two new service tests covering both methods.

Other:
- product_id column upgraded to BigInteger (migration + ORM).
- DEBT-045 → DEBT-051 reference drift in metadata_cache.py and
  docs/modules/statcan.md.
- Test count: 19 (was 16): +1 freshness, +2 product_id mismatch.

+++ backend/migrations/versions/2e15e3f4d1f2_add_cube_metadata_cache.py
@@ -26,7 +26,7 @@ def upgrade() -> None:
         "cube_metadata_cache",
         sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
         sa.Column("cube_id", sa.String(length=50), nullable=False),
-        sa.Column("product_id", sa.Integer(), nullable=False),
+        sa.Column("product_id", sa.BigInteger(), nullable=False),
         sa.Column(
             "dimensions",
             postgresql.JSONB(astext_type=sa.Text()).with_variant(
